change: Rename cvssTransform module
It doesn't contain a class of function. Therefore start the name with a
lower case character and use .js as suffix.
