+++ src/web/components/dashboard/display/cvss/CvssDisplay.jsx
@@ -9,7 +9,7 @@ import FilterTerm from 'gmp/models/filter/filterterm';
 import {isDefined} from 'gmp/utils/identity';
 import React from 'react';
 import BarChart from 'web/components/chart/Bar';
-import transformCvssData from 'web/components/dashboard/display/cvss/CvssTransform';
+import transformCvssData from 'web/components/dashboard/display/cvss/cvssTransform';
 import DataDisplay from 'web/components/dashboard/display/DataDisplay';
 import useGmp from 'web/hooks/useGmp';
 import PropTypes from 'web/utils/PropTypes';
