+++ src/web/components/dashboard/display/cvss/cvssTransform.js
